+++ AccelByte.Sdk/Core/Repository/ITokenRepository.cs
@@ -27,10 +27,14 @@ namespace AccelByte.Sdk.Core.Repository
 
         void StoreToken(LoginType loginType, OauthmodelTokenResponse tokenResponse);
 
+        void StoreToken(LoginType loginType, OauthmodelTokenWithDeviceCookieResponseV3 tokenResponse);
+
         void UpdateToken(OauthmodelTokenResponseV3 tokenResponse);
 
+        void UpdateToken(OauthmodelTokenWithDeviceCookieResponseV3 tokenResponse);
 
-        [Obsolete("Use one with `OauthmodelTokenResponseV3` parameter.", DiagnosticId = "AB_TOKEN_REPO_DEPRECATED_METHOD")]
+
+        [Obsolete("Use one with `OauthmodelTokenResponseV3` or `OauthmodelTokenWithDeviceCookieResponseV3` parameter.", DiagnosticId = "AB_TOKEN_REPO_DEPRECATED_METHOD")]
         void StoreToken(string token);
 
         [Obsolete("Use `Token` property instead.", DiagnosticId = "AB_TOKEN_REPO_DEPRECATED_METHOD")]